requirements.txt
#backports-abc ==0.4
#backports.ssl-match-hostname ==3.4.0.2
flask-sqlalchemy ==2.1
futures==3.0.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.10.4
pandas==0.17.1
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.9.1
scipy==0.17.0
scikit-learn==0.17.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.12
tornado==4.3
Werkzeug==0.11.4
wheel==0.29.0
bokeh==0.11.1
flask-bootstrap==3.3.5.7

